requirements.txt
argon2_cffi
cryptography
tabulate
pyperclip
